backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
google-generativeai==0.8.3